Refactor: Split up main and added type definitions
