+++ components/schemas/ErrorSchema.ts
@@ -1,13 +1,14 @@
-
-import { z} from "npm:@hono/zod-openapi@0.9.5";
+import { z } from "npm:@hono/zod-openapi@0.9.5";
 
 const ErrorSchema = z.object({
-    status: z.number().openapi({ example: 500, }),
-    error: z.string().openapi({ example: "Error fetching data", }),
+  status: z.number().openapi({ example: 500 }),
+  error: z.string().openapi({ example: "Error fetching data" }),
 })
 .openapi({
-    description: "An error response.",
-    example: { status: 400, error: "No results found" },
+  description: "An error response.",
+  example: { status: 400, error: "No results found" },
 });
+
+export type ErrorType = z.infer<typeof ErrorSchema>;
 
 export default ErrorSchema;
